doc_requirements.txt
sphinx-rtd-theme
sphinx-gallery
numpydoc
pillow
ipython
pandas
